+++ custom_components/unifi_insights/api/network/endpoints/__init__.py
@@ -7,9 +7,11 @@ from .clients import ClientsEndpoint
 from .devices import DevicesEndpoint
 from .dns import DNSEndpoint
 from .firewall import FirewallEndpoint
+from .lags import LagsEndpoint
 from .networks import NetworksEndpoint
 from .resources import ResourcesEndpoint
 from .sites import SitesEndpoint
+from .stacks import StacksEndpoint
 from .traffic import TrafficEndpoint
 from .vouchers import VouchersEndpoint
 from .wifi import WifiEndpoint
@@ -20,9 +22,11 @@ __all__ = [
     "DNSEndpoint",
     "DevicesEndpoint",
     "FirewallEndpoint",
+    "LagsEndpoint",
     "NetworksEndpoint",
     "ResourcesEndpoint",
     "SitesEndpoint",
+    "StacksEndpoint",
     "TrafficEndpoint",
     "VouchersEndpoint",
     "WifiEndpoint",